Adding default_vhost true/false param
Any unmatched virtualhost requests are served by httpd to the
virtualhost configuration that was loaded first. Previously the
`priority` param was used to control `apache::vhost` loading order.

This commit adds a `default_vhost` parameter that should be used instead
of the `priority` parameter, and the priority parameter is only used in
specific cases when specific vhost loading order is required (which
currently do not exist, but may).