Updated format and layout
Rearranged subsection headers under Usage and Implementation to more
accurately, clearly and consistently reflect module flow. This
includes: replacing subsection Class: apache, Additional Classes and
Defined Types, and Defines Type: apache::vhost, with  Classes and
Defined Types; updating H-levels of subsections so classes and dts are
h4 and parameters are h5; and moving classes and dts that are not meant
to be used directly by the user to Implementation.

Updated TOC to reflect structure changes.